Resetting the GitHub to after I had added the padding
The GitHub wasn't working and I didn't know what the problem s so I decided to reset it basically by copying the code from when it did work and then pasting into my GitHub. Now I going to try and do it all again but correctly this time. Fingers crossed.