Oops... forgot the line!
Well, so me being the super smart and amazing genius I am, I decided that I would forget the line that split up the different boxes in my blog. However now I have fixed it putting in a solid black 1 pixel line in my website with the code on line 143.