Editing the material inside the masthead
So I didn't actually change anything but the description in my last edit decided to mess up. Changed the name and description at the top of the page and change the avatar picture because this one might actually work better. Just little bit and not any real coding but the changes are on lines 6, 9 and 12.