Updating my table
So this code won't actually come into use until a bit later but better to get it in later. Hoping for future that I will be able to edit the template that we have to build our blogs from. For example getting rid of the second title because it is entirely useless  and untablifying (yes I did just make up my own new word simply because I couldn't be bothered to think of an real word that would fit).