Editing the "blog" and "about buttons in the masthead
I actually made most of the changes in the last update but I forgot to summarise it. In this update  just edited the last few bits and now I hopefully won't have to make anymore changes to the masthead for a while. Yippee! Now I can move on to the rest of the website. By the way, made edits on lines 281 to 285.